Finished the timetable form and enrolments form with small directory changes

+++ Forms/Teaching_materials_form.cs
@@ -8,15 +8,17 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using _401AZ_PROJECT.Classes.Teaching_Materials;  
+using _401AZ_PROJECT.Classes.Teaching_Materials;
+using _401AZ_PROJECT.Classes.Teaching_Materials.TeachingMaterial;
+using _401AZ_PROJECT.Models;
 
 namespace _401AZ_PROJECT
 {
     public partial class Teaching_materials_form : Form
     {
         readonly DataManager DM = new DataManager();
-        readonly TeachingMaterials_Db_Con tm = new TeachingMaterials_Db_Con();
-        readonly FileExtension_Db_Con fe = new FileExtension_Db_Con();
+        readonly TeachingMaterials tm = new TeachingMaterials();
+        readonly FileExtension fe = new FileExtension();
 
         public Teaching_materials_form()
         {
@@ -48,17 +50,19 @@ namespace _401AZ_PROJECT
 
             t.Teacher_id = Int32.Parse(Num_TeacherId.Value.ToString());
 
-            if (DM.ToDataTable(FE.GetFileExtension(FE.File_Extension)).Rows.Count == 0)
+            if (DM.ToDataTable(FE.GetFileExtensionId(FE.File_Extension)).Rows.Count == 0)
             {
                 FE.InsertFileExtension(FE);
             }
-            var file_extension_id = DM.ToDataTable(FE.GetFileExtension(FE.File_Extension)).Rows[0].Field<string>("file_extension_id");
+
+            var file_extension_id = DM.ToDataTable(FE.GetFileExtensionId(FE.File_Extension)).Rows[0].Field<string>("file_extension_id");
             if (Txt_Description.Text.Length < 0)
             {
                 Txt_Description.Text = "No Description";
             }
 
-            t.InsertTeachingMaterial(t.Filename, Convert.ToInt32(file_extension_id), t.Description, rawData, t.Teacher_id);
+            try {t.InsertTeachingMaterial(t.Filename, Convert.ToInt32(file_extension_id), t.Description, rawData, t.Teacher_id);}
+            catch (Exception ex) { MessageBox.Show(ex.ToString()); }
 
 
             Btn_Cancel.PerformClick();
@@ -101,6 +105,7 @@ namespace _401AZ_PROJECT
 
         private void Btn_ChooseFile_Click(object sender, EventArgs e)
         {
+            Fd_Upload.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
             Fd_Upload.ShowDialog();
         }
 
@@ -124,31 +129,37 @@ namespace _401AZ_PROJECT
 
         private void Btn_DeleteRecord_Click(object sender, EventArgs e)
         {
-            int index = Int32.Parse(Dgv_TeachingMaterials.SelectedCells[0].Value.ToString());
-            string caption = "Are you sure you want to delete?";
-            string message = "Do you want to delete the file with the record Id of" + " " + index + " ?";
-            MessageBoxButtons buttons = MessageBoxButtons.YesNo;
-            DialogResult result = MessageBox.Show(message, caption, buttons);
-
-            if (result == DialogResult.Yes)
+            if (Dgv_TeachingMaterials.Rows.Cast<DataGridViewRow>().Any(x => x.Cells.Cast<DataGridViewCell>().Any(c => c.Value != null)))
             {
-                tm.DeleteTeachingMaterial(index);
-                Btn_Refresh.PerformClick();
+                int index = Int32.Parse(Dgv_TeachingMaterials.SelectedCells[0].Value.ToString());
+                string caption = "Are you sure you want to delete?";
+                string message = "Do you want to delete the file with the record Id of" + " " + index + " ?";
+                MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+                DialogResult result = MessageBox.Show(message, caption, buttons);
+
+                if (result == DialogResult.Yes)
+                {
+                    tm.DeleteTeachingMaterial(index);
+                    Btn_Refresh.PerformClick();
+                }
             }
 
         }
 
         private void Btn_Download_Click(object sender, EventArgs e)
         {
-            string fn_no_ext = Dgv_TeachingMaterials.SelectedCells[1].Value.ToString();
-            string fe = Dgv_TeachingMaterials.SelectedCells[2].Value.ToString();
-            
-            sv_Download.Filter = fn_no_ext + fe + "|*" + fe;
-            sv_Download.Title = "Get a file from database";
-            sv_Download.FileName = fn_no_ext + fe;
-            sv_Download.DefaultExt = fe;
-            sv_Download.ShowDialog();
+            if(Dgv_TeachingMaterials.Rows.Cast<DataGridViewRow>().Any(x => x.Cells.Cast<DataGridViewCell>().Any(c => c.Value != null)))
+            {
 
+                string fn_no_ext = Dgv_TeachingMaterials.SelectedCells[1].Value.ToString();
+                string fe = Dgv_TeachingMaterials.SelectedCells[2].Value.ToString();
+                
+                sv_Download.Filter = fn_no_ext + fe + "|*" + fe;
+                sv_Download.Title = "Get a file from database";
+                sv_Download.FileName = fn_no_ext + fe;
+                sv_Download.DefaultExt = fe;
+                sv_Download.ShowDialog();
+            }
         }
 
         private void sv_Download_FileOk(object sender, CancelEventArgs e)
@@ -156,7 +167,7 @@ namespace _401AZ_PROJECT
             int index = Int32.Parse(Dgv_TeachingMaterials.SelectedCells[0].Value.ToString());
             TeachingMaterials downloadData = tm.GetTeachingMaterialsWithFileContentByTeachingId(index);
             try 
-            { 
+            {   
                 File.WriteAllBytes(sv_Download.FileName, downloadData.FileContent);
             }
             catch (Exception ex)
